+++ src/app/dialogs/dialog-wallet-select/dialog-wallet-select.component.ts
@@ -16,13 +16,13 @@ import { Web3Service } from './../../services/web3.service';
   styleUrls: ['./dialog-wallet-select.component.scss']
 })
 export class DialogWalletSelectComponent implements OnInit {
-
   wallets: {
     image: string;
     title: string;
     type: WalletType;
     active?: boolean;
   }[];
+  loggedIn: boolean;
 
   constructor(
     private dialogRef: MatDialogRef<DialogWalletSelectComponent>,
@@ -61,14 +61,24 @@ export class DialogWalletSelectComponent implements OnInit {
    */
   async selectWallet(wallet: WalletType, connected: boolean) {
     if (connected) {
-      await this.web3Service.logout();
-      return this.loadActiveWallet();
+      return;
     }
 
     const loggedIn = await this.web3Service.requestLogin(wallet, true);
     if (loggedIn) {
-      timer(300).subscribe(() => this.dialogRef.close(loggedIn));
+      await timer(300).toPromise();
+      this.dialogRef.close(loggedIn);
     }
+  }
+
+  /**
+   * Logout
+   */
+  async clickLogout() {
+    await this.web3Service.logout();
+    this.loadActiveWallet();
+    this.walletConnectService.disconnect();
+    this.dialogRef.close();
   }
 
   /**
@@ -83,6 +93,8 @@ export class DialogWalletSelectComponent implements OnInit {
    */
   private loadActiveWallet() {
     const loggedIn = this.web3Service.loggedIn;
+    this.loggedIn = loggedIn;
+
     if (!loggedIn) {
       return this.setWalletConnected();
     }
@@ -105,5 +117,4 @@ export class DialogWalletSelectComponent implements OnInit {
       item.active = walletType === item.type || false;
     });
   }
-
 }